requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.34
httpx>=0.28
python-dotenv>=1.0
pydantic-settings>=2.0
